refactor: extract mapping logic from army creation api
